upgrading pointstore
