+++ Java/core/src/main/java/com/amazon/randomcutforest/state/store/PointStoreFloatMapper.java
@@ -18,20 +18,17 @@ package com.amazon.randomcutforest.state.store;
 import static com.amazon.randomcutforest.CommonUtils.checkArgument;
 import static com.amazon.randomcutforest.CommonUtils.checkNotNull;
 
-import java.util.Arrays;
-
 import lombok.Getter;
 import lombok.Setter;
 
 import com.amazon.randomcutforest.config.Precision;
 import com.amazon.randomcutforest.state.IStateMapper;
 import com.amazon.randomcutforest.store.PointStore;
-import com.amazon.randomcutforest.store.PointStoreFloat;
 import com.amazon.randomcutforest.util.ArrayPacking;
 
 @Getter
 @Setter
-public class PointStoreFloatMapper implements IStateMapper<PointStoreFloat, PointStoreState> {
+public class PointStoreFloatMapper implements IStateMapper<PointStore, PointStoreState> {
 
     /**
      * If true, then the arrays are compressed via simple data dependent scheme
@@ -39,7 +36,7 @@ public class PointStoreFloatMapper implements IStateMapper<PointStoreFloat, Poin
     private boolean compressionEnabled = true;
 
     @Override
-    public PointStoreFloat toModel(PointStoreState state, long seed) {
+    public PointStore toModel(PointStoreState state, long seed) {
         checkNotNull(state.getRefCount(), "refCount must not be null");
         checkNotNull(state.getPointData(), "pointData must not be null");
         checkArgument(Precision.valueOf(state.getPrecision()) == Precision.FLOAT_32,
@@ -50,11 +47,10 @@ public class PointStoreFloatMapper implements IStateMapper<PointStoreFloat, Poin
         int startOfFreeSegment = state.getStartOfFreeSegment();
         int[] refCount = ArrayPacking.unpackInts(state.getRefCount(), indexCapacity, state.isCompressed());
         int[] locationList = new int[indexCapacity];
-        Arrays.fill(locationList, PointStore.INFEASIBLE_POINTSTORE_LOCATION);
         int[] tempList = ArrayPacking.unpackInts(state.getLocationList(), state.isCompressed());
         System.arraycopy(tempList, 0, locationList, 0, tempList.length);
 
-        return PointStoreFloat.builder().internalRotationEnabled(state.isRotationEnabled())
+        return PointStore.builder().internalRotationEnabled(state.isRotationEnabled())
                 .internalShinglingEnabled(state.isInternalShinglingEnabled())
                 .dynamicResizingEnabled(state.isDynamicResizingEnabled())
                 .directLocationEnabled(state.isDirectLocationMap()).indexCapacity(indexCapacity)
@@ -65,30 +61,31 @@ public class PointStoreFloatMapper implements IStateMapper<PointStoreFloat, Poin
     }
 
     @Override
-    public PointStoreState toState(PointStoreFloat model) {
+    public PointStoreState toState(PointStore model) {
         model.compact();
         PointStoreState state = new PointStoreState();
         state.setCompressed(compressionEnabled);
         state.setDimensions(model.getDimensions());
         state.setCapacity(model.getCapacity());
         state.setShingleSize(model.getShingleSize());
-        state.setDirectLocationMap(model.isDirectLocationMap());
+        state.setDirectLocationMap(false);
         state.setInternalShinglingEnabled(model.isInternalShinglingEnabled());
         state.setLastTimeStamp(model.getNextSequenceIndex());
         if (model.isInternalShinglingEnabled()) {
             state.setInternalShingle(model.getInternalShingle());
             state.setRotationEnabled(model.isInternalRotationEnabled());
         }
-        state.setDynamicResizingEnabled(model.isDynamicResizingEnabled());
-        if (model.isDynamicResizingEnabled()) {
+        state.setDynamicResizingEnabled(true);
+        if (state.isDynamicResizingEnabled()) {
             state.setCurrentStoreCapacity(model.getCurrentStoreCapacity());
             state.setIndexCapacity(model.getIndexCapacity());
         }
         state.setStartOfFreeSegment(model.getStartOfFreeSegment());
         state.setPrecision(Precision.FLOAT_32.name());
-        int prefix = model.getValidPrefix();
-        state.setRefCount(ArrayPacking.pack(model.getRefCount(), prefix, state.isCompressed()));
-        state.setLocationList(ArrayPacking.pack(model.getLocationList(), prefix, state.isCompressed()));
+        int[] refcount = model.getRefCount();
+        state.setRefCount(ArrayPacking.pack(refcount, refcount.length, state.isCompressed()));
+        int[] locationList = model.getLocationList();
+        state.setLocationList(ArrayPacking.pack(locationList, locationList.length, state.isCompressed()));
         state.setPointData(ArrayPacking.pack(model.getStore(), model.getStartOfFreeSegment()));
         return state;
     }
