+++ Java/core/src/main/java/com/amazon/randomcutforest/state/store/PointStoreState.java
@@ -22,8 +22,8 @@ import lombok.Data;
 /**
  * A class for storing the state of a
  * {@link com.amazon.randomcutforest.store.PointStoreDouble} or a
- * {@link com.amazon.randomcutforest.store.PointStoreFloat}. Depending on which
- * kind of point store was serialized, one of the fields {@code doubleData} or
+ * {@link com.amazon.randomcutforest.store.PointStore}. Depending on which kind
+ * of point store was serialized, one of the fields {@code doubleData} or
  * {@code floatData} will be null.
  */
 @Data
