+++ Java/core/src/main/java/com/amazon/randomcutforest/store/PointStoreDouble.java
@@ -16,38 +16,766 @@
 package com.amazon.randomcutforest.store;
 
 import static com.amazon.randomcutforest.CommonUtils.checkArgument;
+import static com.amazon.randomcutforest.CommonUtils.checkState;
+import static java.lang.Math.max;
 
 import java.util.Arrays;
+import java.util.BitSet;
+import java.util.HashMap;
+import java.util.Optional;
+import java.util.Stack;
+import java.util.Vector;
 
-/**
- * PointStore is a fixed size repository of points, where each point is a float
- * array of a specified length. A PointStore counts references to points that
- * are added, and frees space internally when a given point is no longer in use.
- * The primary use of this store is to enable compression since the points in
- * two different trees do not have to be stored separately.
- *
- * Stored points are referenced by index values which can be used to look up the
- * point values and increment and decrement reference counts. Valid index values
- * are between 0 (inclusive) and capacity (exclusive).
- */
-public class PointStoreDouble extends PointStore<double[], double[]> {
+public class PointStoreDouble implements IPointStore<double[]> {
 
-    public PointStoreDouble(Builder builder) {
-        super(builder);
-        checkArgument(builder.store == null || builder.store.length == currentStoreCapacity * dimensions,
-                " incorrect store length");
-        store = (builder.store == null) ? new double[currentStoreCapacity * dimensions] : builder.store;
+    public static int INFEASIBLE_POINTSTORE_INDEX = -1;
+    public static int INFEASIBLE_LOCN = (int) -2;
+    /**
+     * an index manager to manage free locations
+     */
+    protected IntervalManager indexManager;
+    /**
+     * generic store class
+     */
+    protected double[] store;
+    /**
+     * generic internal shingle, note that input is doubles
+     */
+    protected double[] internalShingle;
+    /**
+     * enable rotation of shingles; use a cyclic buffer instead of sliding window
+     */
+    boolean rotationEnabled;
+    /**
+     * last seen timestamp for internal shingling
+     */
+    protected long nextSequenceIndex;
+    /**
+     * pointers to store locations, this decouples direct addressing and points can
+     * be moved internally
+     */
+    protected int[] locationList;
+    /**
+     * refCount[i] counts of the number of trees that are currently using the point
+     * determined by locationList[i] or (for directLocationMapping) the point at
+     * store[i * dimensions]
+     */
+    protected byte[] refCount;
+
+    protected HashMap<Integer, Integer> refCountMap;
+    /**
+     * first location where new data can be safely copied;
+     */
+    int startOfFreeSegment;
+    /**
+     * overall dimension of the point (after shingling)
+     */
+    int dimensions;
+    /**
+     * shingle size, if known. Setting shingle size = 1 rules out overlapping
+     */
+    int shingleSize;
+    /**
+     * number of original dimensions which are shingled to produce and overall point
+     * dimensions = shingleSize * baseDimensions. However there is a possibility
+     * that even though the data is shingled, we may not choose to use the
+     * overlapping (say for out of order updates).
+     */
+    int baseDimension;
+
+    /**
+     * maximum capacity
+     */
+    int capacity;
+    /**
+     * current capacity of store (number of shingled points)
+     */
+    int currentStoreCapacity;
+
+    /**
+     * enabling internal shingling
+     */
+    boolean internalShinglingEnabled;
+
+    /**
+     * Decrement the reference count for the given index.
+     *
+     * @param index The index value.
+     * @throws IllegalArgumentException if the index value is not valid.
+     * @throws IllegalArgumentException if the current reference count for this
+     *                                  index is non positive.
+     */
+    @Override
+    public int decrementRefCount(int index) {
+        checkArgument(index >= 0 && index < locationList.length, " index not supported by store");
+        checkArgument((refCount[index] & 0xff) > 0, " cannot decrement index");
+        Integer value = refCountMap.remove(index);
+        if (value == null) {
+            if ((refCount[index] & 0xff) == 1) {
+                indexManager.releaseIndex(index);
+                refCount[index] = (byte) 0;
+                locationList[index] = INFEASIBLE_LOCN;
+                return 0;
+            } else {
+                int newVal = (byte) ((refCount[index] & 0xff) - 1);
+                refCount[index] = (byte) newVal;
+                return newVal;
+            }
+        } else {
+            if (value > 1) {
+                refCountMap.put(index, value - 1);
+            }
+            return value - 1 + (refCount[index] & 0xff);
+        }
     }
 
-    public PointStoreDouble(int dimensions, int capacity) {
-        this(new Builder().dimensions(dimensions).shingleSize(1).capacity(capacity).indexCapacity(capacity)
-                .currentStoreCapacity(capacity));
+    /**
+     * takes an index from the index manager and rezises if necessary also adjusts
+     * refCount size to have increment/decrement be seamless
+     *
+     * @return an index from the index manager
+     */
+    int takeIndex() {
+        if (indexManager.isEmpty()) {
+            if (indexManager.getCapacity() < capacity) {
+                int oldCapacity = indexManager.getCapacity();
+                int newCapacity = Math.min(capacity, 1 + (int) Math.floor(1.1 * oldCapacity));
+                indexManager.extendCapacity(newCapacity);
+                refCount = Arrays.copyOf(refCount, newCapacity);
+                int oldlocationLength = locationList.length;
+                locationList = Arrays.copyOf(locationList, newCapacity);
+                for (int i = oldlocationLength; i < newCapacity; i++) {
+                    locationList[i] = INFEASIBLE_LOCN;
+                }
+            } else {
+                throw new IllegalStateException(" index manager in point store is full ");
+            }
+        }
+        return indexManager.takeIndex();
+    }
+
+    protected int getAmountToWrite(double[] tempPoint) {
+        if (checkShingleAlignment(startOfFreeSegment, tempPoint)) {
+            if (!rotationEnabled
+                    || startOfFreeSegment % dimensions == (nextSequenceIndex - 1) * baseDimension % dimensions) {
+                return baseDimension;
+            }
+        } else if (!rotationEnabled) {
+            return dimensions;
+
+        }
+        // the following adds the padding for what exists;
+        // then the padding for the new part; all mod (dimensions)
+        // note that the expression is baseDimension when the condition
+        // startOfFreeSegment % dimensions == (nextSequenceIndex-1)*baseDimension %
+        // dimension
+        // is met
+        return dimensions + (dimensions - startOfFreeSegment % dimensions
+                + (int) ((nextSequenceIndex) * baseDimension) % dimensions) % dimensions;
+    }
+
+    /**
+     * Add a point to the point store and return the index of the stored point.
+     *
+     * @param point       The point being added to the store.
+     * @param sequenceNum sequence number of the point
+     * @return the index value of the stored point.
+     * @throws IllegalArgumentException if the length of the point does not match
+     *                                  the point store's dimensions.
+     * @throws IllegalStateException    if the point store is full.
+     */
+    public int add(double[] point, long sequenceNum) {
+        checkArgument(internalShinglingEnabled || point.length == dimensions,
+                "point.length must be equal to dimensions");
+        checkArgument(!internalShinglingEnabled || point.length == baseDimension,
+                "point.length must be equal to dimensions");
+
+        double[] tempPoint = point;
+        nextSequenceIndex++;
+        if (internalShinglingEnabled) {
+            // rotation is supported via the output and input is unchanged
+            tempPoint = constructShingleInPlace(internalShingle, point, false);
+            if (nextSequenceIndex < shingleSize) {
+                return INFEASIBLE_POINTSTORE_INDEX;
+            }
+        }
+        int nextIndex;
+
+        int amountToWrite = getAmountToWrite(tempPoint);
+
+        if (startOfFreeSegment > currentStoreCapacity * dimensions - amountToWrite) {
+            // try compaction and then resizing
+            compact();
+            // the compaction can change the array contents
+            amountToWrite = getAmountToWrite(tempPoint);
+            if (startOfFreeSegment > currentStoreCapacity * dimensions - amountToWrite) {
+                resizeStore();
+                if (startOfFreeSegment + amountToWrite > currentStoreCapacity * dimensions) {
+                    System.out.println("AA");
+                }
+                checkState(startOfFreeSegment + amountToWrite <= currentStoreCapacity * dimensions, "out of space");
+            }
+        }
+
+        nextIndex = takeIndex();
+
+        locationList[nextIndex] = startOfFreeSegment - dimensions + amountToWrite;
+        if (amountToWrite <= dimensions) {
+            copyPoint(tempPoint, dimensions - amountToWrite, startOfFreeSegment, amountToWrite);
+        } else {
+            copyPoint(tempPoint, 0, startOfFreeSegment + amountToWrite - dimensions, dimensions);
+        }
+        startOfFreeSegment += amountToWrite;
+
+        refCount[nextIndex] = 1;
+        return nextIndex;
+    }
+
+    /**
+     * Increment the reference count for the given index. This operation assumes
+     * that there is currently a point stored at the given index and will throw an
+     * exception if that's not the case.
+     *
+     * @param index The index value.
+     * @throws IllegalArgumentException if the index value is not valid.
+     * @throws IllegalArgumentException if the current reference count for this
+     *                                  index is non positive.
+     */
+    public int incrementRefCount(int index) {
+        checkArgument(index >= 0 && index < locationList.length, " index not supported by store");
+        checkArgument((refCount[index] & 0xff) > 0, " not in use ");
+        Integer value = refCountMap.remove(index);
+        if (value == null) {
+            if ((refCount[index] & 0xff) == 255) {
+                refCountMap.put(index, 1);
+                return 256;
+            } else {
+                int newVal = (byte) ((refCount[index] & 0xff) + 1);
+                refCount[index] = (byte) newVal;
+                return newVal;
+            }
+        } else {
+            refCountMap.put(index, value + 1);
+            return value + 1;
+        }
     }
 
     @Override
+    public int getDimensions() {
+        return dimensions;
+    }
+
+    /**
+     * maximum capacity, in number of points of size dimensions
+     */
+    public int getCapacity() {
+        return capacity;
+    }
+
+    /**
+     * capacity of the indices
+     */
+    public int getIndexCapacity() {
+        return indexManager.getCapacity();
+    }
+
+    /**
+     * used in mapper
+     *
+     * @return gets the shingle size (if known, otherwise is 1)
+     */
+    public int getShingleSize() {
+        return shingleSize;
+    }
+
+    /**
+     * gets the current store capacity in the number of points with dimension many
+     * values
+     *
+     * @return capacity in number of points
+     */
+    public int getCurrentStoreCapacity() {
+        return currentStoreCapacity;
+    }
+
+    /**
+     * used for mappers
+     *
+     * @return the store that stores the values
+     */
+    public double[] getStore() {
+        return store;
+    }
+
+    /**
+     * used for mapper
+     *
+     * @return the array of counts referring to different points
+     */
+    public int[] getRefCount() {
+        int[] newarray = new int[refCount.length];
+        for (int i = 0; i < refCount.length; i++) {
+            newarray[i] = refCount[i] & 0xff;
+            Integer value = refCountMap.get(i);
+            if (value != null) {
+                newarray[i] += value;
+            }
+        }
+        return newarray;
+    }
+
+    /**
+     * useful in mapper to not copy
+     *
+     * @return the length of the prefix
+     */
+    public int getStartOfFreeSegment() {
+        return startOfFreeSegment;
+    }
+
+    /**
+     * used in mapper
+     *
+     * @return if shingling is performed internally
+     */
+    public boolean isInternalShinglingEnabled() {
+        return internalShinglingEnabled;
+    }
+
+    /**
+     * used in mapper and in extrapolation
+     *
+     * @return the last timestamp seen
+     */
+    public long getNextSequenceIndex() {
+        return nextSequenceIndex;
+    }
+
+    /**
+     * used to obtain the most recent shingle seen so far in case of internal
+     * shingling
+     *
+     * @return for internal shingling, returns the last seen shingle
+     */
+    public double[] getInternalShingle() {
+        checkState(internalShinglingEnabled, "internal shingling is not enabled");
+        return copyShingle();
+    }
+
+    /**
+     * The following function eliminates redundant information that builds up in the
+     * point store and shrinks the point store
+     */
+
+    public void compact() {
+
+        Vector<Integer[]> reverseReference = new Vector<>();
+        for (int i = 0; i < locationList.length; i++) {
+            int locn = locationList[i];
+            if (locn < currentStoreCapacity * dimensions && locn >= 0) {
+                reverseReference.add(new Integer[] { locn, i });
+            }
+        }
+        reverseReference.sort((o1, o2) -> o1[0].compareTo(o2[0]));
+        int freshStart = 0;
+        int jStatic = 0;
+        int jDynamic = 0;
+        int jEnd = reverseReference.size();
+        while (jStatic < jEnd) {
+            int blockStart = reverseReference.get(jStatic)[0];
+            int blockEnd = blockStart + dimensions;
+            int initial = 0;
+            if (rotationEnabled) {
+                initial = (dimensions - freshStart + blockStart) % dimensions;
+            }
+            int k = jStatic + 1;
+            jDynamic = jStatic + 1;
+            while (k < jEnd) {
+                int newElem = reverseReference.get(k)[0];
+                if (blockEnd >= newElem) {
+                    k += 1;
+                    jDynamic += 1;
+                    blockEnd = max(blockEnd, newElem + dimensions);
+                } else {
+                    k = jEnd;
+                }
+            }
+
+            // aligning the boundaries
+            for (int i = 0; i < initial; i++) {
+                store[freshStart] = 0;
+                ++freshStart;
+            }
+
+            for (int i = blockStart; i < blockEnd; i++) {
+                store[freshStart] = store[i];
+                assert (!rotationEnabled || freshStart % dimensions == i % dimensions);
+
+                if (jStatic < jEnd) {
+                    int locn = reverseReference.get(jStatic)[0];
+                    if (i == locn) {
+                        int newIdx = reverseReference.get(jStatic)[1];
+                        locationList[newIdx] = freshStart;
+                        jStatic += 1;
+                    }
+                }
+                freshStart += 1;
+            }
+
+            if (jStatic != jDynamic) {
+                throw new IllegalStateException("There is discepancy in indices");
+            }
+        }
+        startOfFreeSegment = freshStart;
+    }
+
+    /**
+     * returns the number of copies of a point
+     *
+     * @param i index of a point
+     * @return number of copies of the point managed by the store
+     */
+    public int getRefCount(int i) {
+        int val = refCount[i] & 0xff;
+        Integer value = refCountMap.get(i);
+        if (value != null) {
+            val += value;
+        }
+        return val;
+    }
+
+    @Override
+    public boolean isInternalRotationEnabled() {
+        return rotationEnabled;
+    }
+
+    public int[] getLocationList() {
+        return locationList;
+    }
+
+    /**
+     *
+     * @return the number of indices stored
+     */
+    public int size() {
+        int count = 0;
+        for (int i = 0; i < locationList.length; i++) {
+            if (locationList[i] != INFEASIBLE_LOCN) {
+                ++count;
+            }
+        }
+        return count;
+    }
+
+    /**
+     * transforms a point to a shingled point if internal shingling is turned on
+     *
+     * @param point new input values
+     * @return shingled point
+     */
+    @Override
+    public double[] transformToShingledPoint(double[] point) {
+        checkArgument(internalShinglingEnabled, " only allowed for internal shingling");
+        checkArgument(point.length == baseDimension, " incorrect length");
+        return constructShingleInPlace(copyShingle(), point, rotationEnabled);
+    }
+
+    private double[] copyShingle() {
+        if (!rotationEnabled) {
+            return Arrays.copyOf(internalShingle, dimensions);
+        } else {
+            double[] answer = new double[dimensions];
+            int offset = (int) (nextSequenceIndex * baseDimension);
+            for (int i = 0; i < dimensions; i++) {
+                answer[(offset + i) % dimensions] = internalShingle[i];
+            }
+            return answer;
+        }
+    }
+
+    /**
+     * the following function is used to update the shingle in place; it can be used
+     * to produce new copies as well
+     *
+     * @param target the array containing the shingled point
+     * @param point  the new values
+     * @return the array which now contains the updated shingle
+     */
+    protected double[] constructShingleInPlace(double[] target, double[] point, boolean rotationEnabled) {
+        if (!rotationEnabled) {
+            for (int i = 0; i < dimensions - baseDimension; i++) {
+                target[i] = target[i + baseDimension];
+            }
+            for (int i = 0; i < baseDimension; i++) {
+                target[dimensions - baseDimension + i] = (point[i] == 0.0) ? 0.0 : point[i];
+            }
+        } else {
+            int offset = ((int) (nextSequenceIndex * baseDimension) % dimensions);
+            for (int i = 0; i < baseDimension; i++) {
+                target[offset + i] = (point[i] == 0.0) ? 0.0 : point[i];
+            }
+        }
+        return target;
+    }
+
+    /**
+     * for extrapolation and imputation, in presence of internal shingling we need
+     * to update the list of missing values from the space of the input dimensions
+     * to the shingled dimensions
+     *
+     * @param indexList list of missing values in the input point
+     * @return list of missing values in the shingled point
+     */
+    @Override
+    public int[] transformIndices(int[] indexList) {
+        checkArgument(internalShinglingEnabled, " only allowed for internal shingling");
+        checkArgument(indexList.length <= baseDimension, " incorrect length");
+        int[] results = Arrays.copyOf(indexList, indexList.length);
+        if (!rotationEnabled) {
+            for (int i = 0; i < indexList.length; i++) {
+                checkArgument(results[i] < baseDimension, "incorrect index");
+                results[i] += dimensions - baseDimension;
+            }
+        } else {
+            int offset = ((int) (nextSequenceIndex * baseDimension) % dimensions);
+            for (int i = 0; i < indexList.length; i++) {
+                checkArgument(results[i] < baseDimension, "incorrect index");
+                results[i] = (results[i] + offset) % dimensions;
+            }
+        }
+        return results;
+    }
+
+    /**
+     * a builder
+     */
+
+    public static class Builder<T extends Builder<T>> {
+
+        // We use Optional types for optional primitive fields when it doesn't make
+        // sense to use a constant default.
+
+        protected int dimensions;
+        protected int shingleSize = 1;
+        protected int baseDimension;
+        protected boolean internalRotationEnabled = false;
+        protected boolean internalShinglingEnabled = false;
+        protected int capacity;
+        protected Optional<Integer> initialPointStoreSize = Optional.empty();
+        protected int currentStoreCapacity = 0;
+        protected int indexCapacity = 0;
+        protected double[] store = null;
+        protected double[] knownShingle = null;
+        protected int[] locationList = null;
+        protected int[] refCount = null;
+        protected long nextTimeStamp = 0;
+        protected int startOfFreeSegment = 0;
+
+        // dimension of the points being stored
+        public T dimensions(int dimensions) {
+            this.dimensions = dimensions;
+            return (T) this;
+        }
+
+        // maximum number of points in the store
+        public T capacity(int capacity) {
+            this.capacity = capacity;
+            return (T) this;
+        }
+
+        // initial size of the pointstore, dynamicResizing must be on
+        // and value cannot exceed capacity
+        public T initialSize(int initialPointStoreSize) {
+            this.initialPointStoreSize = Optional.of(initialPointStoreSize);
+            return (T) this;
+        }
+
+        // shingleSize for opportunistic compression
+        public T shingleSize(int shingleSize) {
+            this.shingleSize = shingleSize;
+            return (T) this;
+        }
+
+        // is internal shingling enabled
+        public T internalShinglingEnabled(boolean internalShinglingEnabled) {
+            this.internalShinglingEnabled = internalShinglingEnabled;
+            return (T) this;
+        }
+
+        @Deprecated
+        public T directLocationEnabled(boolean value) {
+            return (T) this;
+        }
+
+        @Deprecated
+        public T dynamicResizingEnabled(boolean value) {
+            return (T) this;
+        }
+
+        // are shingles rotated
+        public T internalRotationEnabled(boolean internalRotationEnabled) {
+            this.internalRotationEnabled = internalRotationEnabled;
+            return (T) this;
+        }
+
+        // the size of the array storing the specific points
+        // this is used for serialization
+        public T currentStoreCapacity(int currentStoreCapacity) {
+            this.currentStoreCapacity = currentStoreCapacity;
+            return (T) this;
+        }
+
+        // the size of the pointset being tracked
+        // this is used for serialization
+        public T indexCapacity(int indexCapacity) {
+            this.indexCapacity = indexCapacity;
+            return (T) this;
+        }
+
+        // last known shingle, if internalshingle is on
+        // this shingle is not rotated
+        // this is used for serialization
+        public T knownShingle(double[] knownShingle) {
+            this.knownShingle = knownShingle;
+            return (T) this;
+        }
+
+        // count of the points being tracked
+        // used for serialization
+        public T refCount(int[] refCount) {
+            this.refCount = refCount;
+            return (T) this;
+        }
+
+        // location of the points being tracked, if not directmapped
+        // used for serialization
+        public T locationList(int[] locationList) {
+            this.locationList = locationList;
+            return (T) this;
+        }
+
+        public T store(double[] store) {
+            this.store = store;
+            return (T) this;
+        }
+
+        // location of where points can be written
+        // used for serialization
+        public T startOfFreeSegment(int startOfFreeSegment) {
+            this.startOfFreeSegment = startOfFreeSegment;
+            return (T) this;
+        }
+
+        // the next timeStamp to accept
+        // used for serialization
+        public T nextTimeStamp(long nextTimeStamp) {
+            this.nextTimeStamp = nextTimeStamp;
+            return (T) this;
+        }
+
+        public PointStoreDouble build() {
+            return new PointStoreDouble(this);
+        }
+    }
+
+    public PointStoreDouble(int dimensions, int capacity) {
+        this(PointStoreDouble.builder().capacity(capacity).dimensions(dimensions).shingleSize(1).initialSize(capacity));
+    }
+
+    public PointStoreDouble(Builder builder) {
+        checkArgument(builder.dimensions > 0, "dimensions must be greater than 0");
+        checkArgument(builder.capacity > 0, "capacity must be greater than 0");
+        checkArgument(builder.shingleSize == 1 || builder.dimensions == builder.shingleSize
+                || builder.dimensions % builder.shingleSize == 0, "incorrect use of shingle size");
+        /**
+         * the following checks are due to mappers (kept for future)
+         */
+        if (builder.refCount != null || builder.locationList != null || builder.knownShingle != null) {
+            checkArgument(builder.refCount != null, "reference count must be present");
+            checkArgument(builder.locationList != null, "location list must be present");
+            checkArgument(builder.refCount.length == builder.indexCapacity, "incorrect reference count length");
+            // following may change if IndexManager is dynamically resized as well
+            checkArgument(builder.locationList.length == builder.indexCapacity, " incorrect length of locations");
+            checkArgument(
+                    builder.knownShingle == null
+                            || builder.internalShinglingEnabled && builder.knownShingle.length == builder.dimensions,
+                    "incorrect shingling information");
+        }
+
+        this.shingleSize = builder.shingleSize;
+        this.dimensions = builder.dimensions;
+        this.internalShinglingEnabled = builder.internalShinglingEnabled;
+        this.rotationEnabled = builder.internalRotationEnabled;
+        this.baseDimension = this.dimensions / this.shingleSize;
+        this.capacity = builder.capacity;
+        this.refCountMap = new HashMap<>();
+
+        if (builder.refCount == null) {
+            int size = (int) builder.initialPointStoreSize.orElse(builder.capacity);
+            currentStoreCapacity = size;
+            this.indexManager = new IntervalManager(size);
+            startOfFreeSegment = 0;
+            refCount = new byte[size];
+            if (internalShinglingEnabled) {
+                nextSequenceIndex = 0;
+                internalShingle = new double[dimensions];
+            }
+            store = new double[currentStoreCapacity * dimensions];
+            locationList = new int[currentStoreCapacity];
+            Arrays.fill(locationList, INFEASIBLE_LOCN);
+        } else {
+            this.refCount = new byte[builder.refCount.length];
+            for (int i = 0; i < refCount.length; i++) {
+                if (builder.refCount[i] >= 0 && builder.refCount[i] <= 255) {
+                    refCount[i] = (byte) builder.refCount[i];
+                } else if (builder.refCount[i] > 255) {
+                    refCount[i] = (byte) 255;
+                    refCountMap.put(i, builder.refCount[i] - 255);
+                }
+            }
+            this.startOfFreeSegment = builder.startOfFreeSegment;
+            this.nextSequenceIndex = builder.nextTimeStamp;
+            this.currentStoreCapacity = builder.currentStoreCapacity;
+            if (internalShinglingEnabled) {
+                this.internalShingle = (builder.knownShingle != null) ? Arrays.copyOf(builder.knownShingle, dimensions)
+                        : new double[dimensions];
+            }
+            BitSet bits = new BitSet(refCount.length);
+            for (int i = 0; i < refCount.length; i++) {
+                if ((refCount[i] & 0xff) > 0) {
+                    bits.set(i);
+                }
+            }
+            int first = bits.nextClearBit(0);
+            Stack<int[]> stack = new Stack<>();
+
+            while (first < refCount.length) {
+                int last = bits.nextSetBit(first) - 1;
+                if (last >= first) {
+                    stack.push(new int[] { first, last });
+                    first = last + 1;
+                } else { // we do not all distiction between all full and all empty
+                    if (first < refCount.length - 1) {
+                        if (bits.nextClearBit(first + 1) == first + 1) {
+                            stack.push(new int[] { first, refCount.length - 1 });
+                        } else {
+                            stack.push(new int[] { first, first });
+                        }
+                    } else {
+                        stack.push(new int[] { refCount.length - 1, refCount.length - 1 });
+                    }
+                    break;
+                }
+            }
+            indexManager = new IntervalManager(stack, builder.indexCapacity);
+            store = (builder.store == null) ? new double[currentStoreCapacity * dimensions] : builder.store;
+            this.locationList = builder.locationList;
+        }
+    }
+
     void resizeStore() {
         int maxCapacity = (rotationEnabled) ? 2 * capacity : capacity;
-        int newCapacity = Math.min(2 * currentStoreCapacity, maxCapacity);
+        int newCapacity = (int) Math.floor(Math.min(1.1 * currentStoreCapacity, maxCapacity));
         if (newCapacity > currentStoreCapacity) {
             double[] newStore = new double[newCapacity * dimensions];
             System.arraycopy(store, 0, newStore, 0, currentStoreCapacity * dimensions);
@@ -56,18 +784,18 @@ public class PointStoreDouble extends PointStore<double[], double[]> {
         }
     }
 
-    @Override
     boolean checkShingleAlignment(int location, double[] point) {
         boolean test = (location - dimensions + baseDimension >= 0);
         for (int i = 0; i < dimensions - baseDimension && test; i++) {
-            test = (point[i] == store[location - dimensions + baseDimension + i]);
+            test = (((float) point[i]) == store[location - dimensions + baseDimension + i]);
         }
         return test;
     }
 
-    @Override
     void copyPoint(double[] point, int src, int location, int length) {
-        System.arraycopy(point, src, store, location, length);
+        for (int i = 0; i < length; i++) {
+            store[location + i] = point[src + i];
+        }
     }
 
     /**
@@ -87,9 +815,10 @@ public class PointStoreDouble extends PointStore<double[], double[]> {
 
     @Override
     public boolean pointEquals(int index, double[] point) {
-        indexManager.checkValidIndex(index);
+        checkArgument(index >= 0 && index < locationList.length, " index not supported by store");
         checkArgument(point.length == dimensions, "point.length must be equal to dimensions");
         int address = locationList[index];
+        checkArgument(locationList[index] != INFEASIBLE_LOCN, " invalid point");
         if (!rotationEnabled) {
             for (int j = 0; j < dimensions; j++) {
                 if (point[j] != store[j + address]) {
@@ -103,6 +832,7 @@ public class PointStoreDouble extends PointStore<double[], double[]> {
                 }
             }
         }
+
         return true;
     }
 
@@ -118,8 +848,9 @@ public class PointStoreDouble extends PointStore<double[], double[]> {
      */
     @Override
     public double[] get(int index) {
-        indexManager.checkValidIndex(index);
+        checkArgument(index >= 0 && index < locationList.length, " index not supported by store");
         int address = locationList[index];
+        checkArgument(locationList[index] != INFEASIBLE_LOCN, " invalid point");
         if (!rotationEnabled) {
             return Arrays.copyOfRange(store, address, address + dimensions);
         } else {
@@ -131,7 +862,6 @@ public class PointStoreDouble extends PointStore<double[], double[]> {
         }
     }
 
-    // same as get; allows a multiplier to enable convex combinations
     public double[] getScaledPoint(int index, double factor) {
         double[] answer = get(index);
         for (int i = 0; i < dimensions; i++) {
@@ -140,18 +870,14 @@ public class PointStoreDouble extends PointStore<double[], double[]> {
         return answer;
     }
 
-    @Override
     public String toString(int index) {
         return Arrays.toString(get(index));
     }
 
-    @Override
     void copyTo(int dest, int source, int length) {
         // validateInternalState(dest <= source, "error");
-        if (dest < source) {
-            for (int i = 0; i < length; i++) {
-                store[dest + i] = store[source + i];
-            }
+        for (int i = 0; i < length; i++) {
+            store[dest + i] = store[source + i];
         }
     }
 
@@ -159,16 +885,4 @@ public class PointStoreDouble extends PointStore<double[], double[]> {
         return new Builder();
     }
 
-    public static class Builder extends PointStore.Builder<Builder> {
-        private double[] store = null;
-
-        public Builder store(double[] store) {
-            this.store = store;
-            return this;
-        }
-
-        public PointStoreDouble build() {
-            return new PointStoreDouble(this);
-        }
-    }
 }
